refactor: move router from main to app
