+++ src/App.tsx
@@ -1,57 +1,15 @@
-import { useState } from 'react';
-import Content from './components/Content/Content';
-import Search from './components/Search/Search';
-import ListSettings from './components/ListSettings/ListSettings';
-import { useSearchParams } from 'react-router-dom';
+import { RouterProvider, createBrowserRouter } from 'react-router-dom';
+import MainPage from './pages/MainPage';
 
-const DEFAULT_PAGE_SIZE = 6;
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <MainPage />,
+  },
+]);
 
 function App() {
-  const [query, setQuery] = useState(localStorage.getItem('searchQuery') || '');
-  const [pageSize, setPageSize] = useState(
-    localStorage.getItem('pageSize') || DEFAULT_PAGE_SIZE.toString()
-  );
-
-  const [params, setParams] = useSearchParams();
-
-  const handleSearch = (data: string) => {
-    setQuery(data);
-    localStorage.setItem('searchQuery', data);
-  };
-
-  const handleListSettings = (data: number) => {
-    setPageSize(data.toString());
-    localStorage.setItem('pageSize', data.toString());
-    setParams((prev) => {
-      prev.set('page', '1');
-      return prev;
-    });
-  };
-
-  const handlePage = (data: number) => {
-    setParams((prev) => {
-      prev.set('page', data.toString());
-      return prev;
-    });
-  };
-
-  return (
-    <>
-      <Search onSubmit={handleSearch} query={query} />
-      <br />
-      <ListSettings
-        onPageSizeChange={handleListSettings}
-        pageSize={Number(pageSize)}
-      />
-      <hr />
-      <Content
-        query={query}
-        onPageChange={handlePage}
-        pageSize={Number(pageSize)}
-        page={Number(params.get('page')) || 1}
-      />
-    </>
-  );
+  return <RouterProvider router={router} />;
 }
 
 export default App;
