+++ src/main.tsx
@@ -1,19 +1,11 @@
 import ReactDOM from 'react-dom/client';
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import App from './App.tsx';
 import './index.css';
 import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary.tsx';
 import Fallback from './components/Fallback/Fallback.tsx';
 
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <App />,
-  },
-]);
-
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <ErrorBoundary fallback={<Fallback />}>
-    <RouterProvider router={router} />
+    <App />
   </ErrorBoundary>
 );
